requirements.txt
pandas==1.3.1
scikit-learn==0.23.2
streamlit==0.86.0
pillow~=8.3.1
matplotlib~=3.4.3
seaborn~=0.11.2